Create GEOCACHELOGBOOK
Assignment 6 - Geocaching Log Book 
Due Monday by 11:59pm Points 20 Submitting a text entry box or a file upload
Geocaching is an outdoor recreational activity, in which participants use a Global Positioning System (GPS) receiver or mobile device and other navigational techniques to hide and seek containers, called "geocaches" or "caches", at specific locations marked by coordinates all over the world. As of 2021 there were over a million active players in the United States.

Geocaching can be considered a Location-based game. A typical cache is a small waterproof container containing a logbook and sometimes a pen or pencil. The geocacher signs the log with their established code name and dates it, in order to prove that they found the cache. After signing the log, the cache must be placed back exactly where the person found it.

Check out the wikipedia page here: https://en.wikipedia.org/wiki/GeocachingLinks to an external site.

For this assignment, you will replace the paper username log with a program that will store the usernames of players who find the geocache.  Write a program that will add a player's username to a geocache log.  This will be part of a larger geocaching program, so only a few functions are required.  You can assume that other programmers will complete the other parts of the geocaching application.

The main Function
In the main function (5 points), perform the following tasks:

ask the for username and store it in a variable
create an array of size 10  and fill each position with the string "empty"
ask the user to identify the item found.  If the item is equal to "original can of beans", then call the getSize function. Store the return value in a variable called arraySize. 
If the arraySize variable is not -1, call the addUsername function and pass the array, the arraySize variable,  and username variable as arguments.
The getSize Function
The getSize function (5 points)  will return the number of array positions that hold usernames.  Using a loop, count the number of array positions that do not hold the value "empty".   If there are no positions that contain the string "empty", then the array is full, so display the error message "original can of beans destroyed by a lawn mower" and return -1, otherwise, return the count.

The addUsername Function
The addUsername function (5 points) will add a username to an array.  This function requires three parameters:

username
array
arraySize variable
This function should only have one line of code.  Add the username to the array in the arraySize position.

The displayUsernames Function
The displayUsernames function (5 points) should use a loop to display all array positions that do not contain the word "empty".