Create 11.2.5.10pm
how do i call getSize and have it print out the array in ...#include <iostream>
#include <string>
using namespace std;
int getSize(string array[], int arraySize) [
 int count = 0;
 for (int i = 0; i < arraySize; i++) {
  if (array[i] != "empty") {
   count ++;
 }
}
int main() { 
string array[10];
for (int i = 0; i < 10; i++) {
 array[i] = "empty";
}
The problem with your code is that you're not calling getSize(). To call getSize(), you would need to do the following:

C++
int arraySize = 10;
string array[arraySize];

// ...

int size = getSize(array, arraySize);
Use code with caution. Learn more
Then, you could print out the size of the array like this:

C++
cout << "The size of the array is: " << size << endl;
Use code with caution. Learn more
You also need to initialize the array before you call getSize(). To do this, you would need to do the following:

C++
string array[arraySize];

for (int i = 0; i < arraySize; i++) {
 array[i] = "empty";
}
Use code with caution. Learn more
Once you have initialized the array, you can call getSize() and print out the size of the array.

profile picture
did i not already indicate that the size of the array was array[10]
You did indicate that the size of the array was 10, but you did not initialize the array before you called getSize(). To initialize the array, you would need to do the following:

C++
string array[10];

for (int i = 0; i < 10; i++) {
 array[i] = "empty";
}
Use code with caution. Learn more
Once you have initialized the array, you can call getSize() and print out the size of the array.

Here is an example of how to use getSize() to print out the size of an array:

C++
#include <iostream>
#include <string>
using namespace std;

int getSize(string array[], int arraySize) {
 int count = 0;
 for (int i = 0; i < arraySize; i++) {
  if (array[i] != "empty") {
   count ++;
 }
}
 return count;
}

int main() {
 string array[10];

 // Initialize the array
 for (int i = 0; i < 10; i++) {
   array[i] = "empty";
 }

 // Get the size of the array
 int size = getSize(array, 10);

 // Print out the size of the array
 cout << "The size of the array is: " << size << endl;

 return 0;
}
Use code with caution. Learn more
Output:

The size of the array is: 0
profile picture
should not the output be the size of the array is: 10
View other drafts



















The size of the array is 0 because the array is initialized with all empty strings. The getSize() function only counts the number of strings in the array that are not empty.

If you want to print out the size of the array, you can use the sizeof() operator. The sizeof() operator returns the size of a type in bytes. For example, to print out the size of the array, you would do the following:

C++
int size = sizeof(array);

cout << "The size of the array is: " << size << endl;
Use code with caution. Learn more
The output of this code would be:

The size of the array is: 40
This is because the array is a variable-length array, and the sizeof() operator returns the size of the array in bytes, including the size of the pointer to the array.